Moved search state and methods down to the add-project-agents component
